+++ helpers/safeTest.js
@@ -1,0 +1,9 @@
+async function safeTest(testFunc, name) {
+  try {
+    await testFunc();
+  } catch (e) {
+    console.error(`❌ Error in ${name}:`, e.message);
+  }
+}
+
+module.exports = safeTest;
