+++ tests/buttonVisibilityTest.js
@@ -1,0 +1,8 @@
+module.exports = async function testButtonVisibility(driver) {
+  try {
+    const btns = await driver.findElements({ css: 'button, input[type="button"]' });
+    console.log(`🔘 Found ${btns.length} buttons`);
+  } catch {
+    console.log('ℹ️ No buttons found');
+  }
+};
